refactor: flatten module structure
