--- honeycomb-core/src/cmap/mod.rs
@@ -1,27 +1,0 @@
-//! Module short description
-//!
-//! Should you interact with this module directly?
-//!
-//! Content description if needed
-
-// ------ MODULE DECLARATIONS
-
-pub mod dart;
-pub mod embed;
-pub mod map;
-
-// ------ IMPORTS
-
-// ------ CONTENT
-
-// ------ TESTS
-
-#[cfg(test)]
-mod tests {
-    //use super::*;
-
-    #[test]
-    fn some_test() {
-        assert_eq!(1, 1);
-    }
-}
